+++ include/View/egnine.h
@@ -100,7 +100,7 @@ class GraphicsEngine {
 	void record_draw_command(vk::CommandBuffer commandBuffer, uint32_t imageIndex);
 
 	void prepare_scene(vk::CommandBuffer commandBuffer);
-	void render_objects(vk::CommandBuffer commandBuffer, int objectType, uint32_t& startInstance, uint32_t instanceCount);
+	void render_objects(vk::CommandBuffer commandBuffer, uint64_t objectType, uint32_t& startInstance, uint32_t instanceCount);
 
 	void prepare_frame(uint32_t imageIndex, Scene* scene, float deltaTime, Camera::Camera camera);
